requirements.txt
flask
requests
python-dotenv
supabase
gunicorn
